refactor: domainのimport周りを整理
