+++ api/src/application/film_record/query/film_record_query_application.py
@@ -1,14 +1,16 @@
+# 標準ライブラリ
 from logging import getLogger
 
+# 独自ライブラリ
 from src.application.film_record.query.film_record_query_domain_service import ImplFilmRecordQueryDomainService
 from src.application.film_record.query.interface.film_record_query_application_interface import (
     IFilmRecordQueryApplication,
 )
 from src.application.film_record.query.interface.film_record_query_domain_service import IFilmRecordQueryDomainService
 from src.domain.film_record import FilmRecordEntity, FilmRecordIdObject, IFilmRecordRepository
-from src.domain.film_record.film.genre.film_genre_repository import IFilmGenreRepository
+from src.domain.film_record.film.genre import IFilmGenreRepository
 from src.domain.film_record.film.series.film_series_repository import IFilmSeriesRepository
-from src.domain.film_record.watch_history.watch_history_repository import IFilmWatchHistoryRepository
+from src.domain.film_record.watch_history import IFilmWatchHistoryRepository
 from src.domain.film_record.watch_medium.watch_medium_repository import IWatchMediumRepository
