+++ api/src/application/film_record/query/interface/film_record_query_application_interface.py
@@ -1,9 +1,11 @@
+# 標準ライブラリ
 from abc import ABC, abstractmethod
 
+# 独自ライブラリ
 from src.domain.film_record import FilmRecordEntity, IFilmRecordRepository
-from src.domain.film_record.film.genre.film_genre_repository import IFilmGenreRepository
+from src.domain.film_record.film.genre import IFilmGenreRepository
 from src.domain.film_record.film.series.film_series_repository import IFilmSeriesRepository
-from src.domain.film_record.watch_history.watch_history_repository import IFilmWatchHistoryRepository
+from src.domain.film_record.watch_history import IFilmWatchHistoryRepository
 from src.domain.film_record.watch_medium.watch_medium_repository import IWatchMediumRepository
