+++ api/src/domain/film_record/__init__.py
@@ -1,4 +1,3 @@
-# flake8: noqa
 from .film_record_entity import FilmRecordEntity
 from .film_record_id_object import FilmRecordIdObject
 from .film_record_repository import IFilmRecordRepository
