+++ api/src/domain/film_record/appreciation/__init__.py
@@ -1,4 +1,3 @@
-# flake8: noqa
 from .appreciation_status_enum import AppreciationStatusEnum
 from .film_appreciation_entity import FilmAppreciationEntity
 from .film_appreciation_id_object import FilmAppreciationIdObject
