+++ api/src/domain/film_record/appreciation/appreciation_status_enum.py
@@ -1,3 +1,4 @@
+# 標準ライブラリ
 from enum import Enum
