+++ api/src/domain/film_record/dto/__init__.py
@@ -1,0 +1,1 @@
+from .film_record_dto import FilmRecordDto
