+++ api/src/domain/film_record/film/__init__.py
@@ -1,0 +1,2 @@
+from .film_entity import FilmEntity
+from .tmdb_id_object import TmdbIdObject
