+++ api/src/domain/film_record/film/genre/__init__.py
@@ -1,0 +1,3 @@
+from .film_genre_dto import FilmGenreDTO
+from .film_genre_enum import FilmGenreEnum
+from .film_genre_repository import IFilmGenreRepository
