+++ api/src/domain/film_record/film/genre/film_genre_enum.py
@@ -1,3 +1,4 @@
+# 標準ライブラリ
 from enum import Enum
