+++ api/src/domain/film_record/film/poster/__init__.py
@@ -1,0 +1,1 @@
+from .film_poster_object import FilmPosterObject
