+++ api/src/domain/film_record/film/poster/film_poster_object.py
@@ -1,3 +1,4 @@
+# 標準ライブラリ
 import re
 from dataclasses import dataclass
