+++ api/src/domain/film_record/film/series/__init__.py
@@ -1,0 +1,2 @@
+from .film_series_object import FilmSeriesObject
+from .film_series_repository import IFilmSeriesRepository
