+++ api/src/domain/film_record/film/series/film_series_repository.py
@@ -1,3 +1,4 @@
+# 標準ライブラリ
 from abc import ABC, abstractmethod
