+++ api/src/domain/film_record/film_record_id_object.py
@@ -1,3 +1,4 @@
+# 標準ライブラリ
 from dataclasses import dataclass
