+++ api/src/domain/film_record/watch_medium/watch_medium_enum.py
@@ -1,3 +1,4 @@
+# 標準ライブラリ
 from enum import Enum
