+++ api/src/domain/film_record/watch_medium/watch_medium_repository.py
@@ -1,3 +1,4 @@
+# 標準ライブラリ
 from abc import ABC, abstractmethod
