+++ api/src/infrastructure/film_record/watch_history/inmemory/inmemory_watch_history_repository.py
@@ -1,8 +1,12 @@
+# 標準ライブラリ
 import os
 from logging import getLogger
 
+# 外部ライブラリ
 from notion_client import Client
-from src.domain.film_record.watch_history.watch_history_repository import IFilmWatchHistoryRepository
+
+# 独自ライブラリ
+from src.domain.film_record.watch_history import IFilmWatchHistoryRepository
 
 logger = getLogger(__name__)
