+++ api/tests/domain/film_record/test_film_record_entity.py
@@ -1,13 +1,16 @@
+# 標準ライブラリ
 from datetime import date
 
+# 外部ライブラリ
 import pytest
+
+# 独自ライブラリ
 from src.domain.film_record import FilmRecordEntity, FilmRecordIdObject
 from src.domain.film_record.appreciation import AppreciationStatusEnum, FilmAppreciationEntity, FilmAppreciationIdObject
-from src.domain.film_record.film.film_entity import FilmEntity
-from src.domain.film_record.film.genre.film_genre_enum import FilmGenreEnum
-from src.domain.film_record.film.poster.film_poster_object import FilmPosterObject
-from src.domain.film_record.film.series.film_series_object import FilmSeriesObject
-from src.domain.film_record.film.tmdb_id_object import TmdbIdObject
+from src.domain.film_record.film import FilmEntity, TmdbIdObject
+from src.domain.film_record.film.genre import FilmGenreEnum
+from src.domain.film_record.film.poster import FilmPosterObject
+from src.domain.film_record.film.series import FilmSeriesObject
 from src.domain.film_record.watch_medium.watch_medium_enum import WatchMediumEnum
